Create nmapsynscan
-sS: Specifies a SYN Scan, which is faster than a TCP Connect Scan.
-sV: Enables version detection to determine the version of services running on the target.
--version-intensity 5: Sets the intensity of version detection to maximum.
-p 1-65535: Specifies the range of ports to scan.
-oX: Specifies the output format as an XML file.
scan_results.xml: Specifies the output file name as scan_results.xml.
target_ip_address: Specifies the IP address of the target to scan.
This scan will perform a SYN Scan on all ports in the range of 1-65535 and will attempt to identify the versions of the services running on those ports with maximum intensity. The output will be saved in an XML file named scan_results.xml.